requirements.txt
flask == 1.1.2
flask-sqlalchemy == 2.4.4
flask-migrate == 2.5.3
flask-marshmallow == 0.14.0
marshmallow-sqlalchemy == 0.24.1
alembic == 1.4.3
psycopg2 == 2.8.6
PyJWT == 2.0.1